+++ NEWS.md
@@ -5,16 +5,21 @@ N.B. All statistical analysis in `ggstatsplot` is carried out in
 read the `NEWS` for that package:
 <https://indrajeetpatil.github.io/statsExpressions/news/index.html>
 
-There is finally a publication for `ggstatsplot` package! 🎉
-
-<https://joss.theoj.org/papers/10.21105/joss.03167>
-
 NEW FUNCTIONS
 
   - Adds `extract_stats` function to extract dataframes containing statistical
     details.
 
 MAJOR CHANGES
+
+  - There is finally a publication for `ggstatsplot` package!
+    <https://joss.theoj.org/papers/10.21105/joss.03167>
+
+  - The `ggcoefstats` function defaults to `NULL` for `xlab` and `ylab`
+    arguments, which lets users change these labels if they wish to do so.
+    Additionally, the x-axis label, if not specified, now defaults to
+    `"estimate"`. Whether this estimate corresponds to regression coefficient or
+    effect size like partial eta-squared should be clear from the label itself.
 
   - To reduce the dependency load, `ggcorrplot` moves from `Imports` to
     `Suggests`.
@@ -38,6 +43,11 @@ MINOR CHANGES
 
   - Minor change to `ggcorrmat` legend title - content in parentheses is now
     shown outside of it.
+
+BUG FIXES
+
+  - `ggcoefstats` didn't work when statistic for the given model was
+    chi-squared. This has been fixed.
 
 # ggstatsplot 0.7.2